Install a public util header and bump API version.
We are about to open-source the event recorder API since it is used in
mutter, and we want to comply with mutter's GPL license. Rename
get_current_time to emtr_util_get_current_time and put it in its own
utility file (eosmetrics/emtr-util.h) in order to make it accessible
from other repositories. get_uuid_builder doesn't merit sharing, so
move it to eosmetrics/emtr-event-recorder.c since that's the only
file that uses it. Bumped the minor version number from 0 to 2 since we
want to be able to require the version of the event recorder API that
installs eosmetrics/emtr-util.h. Make minor stylistic improvements in
the process of preparing the repository to be open-sourced.

[endlessm/eos-sdk#2043]

(Kurt & Philip)