Bump API version from 0.0.0 to 0.2.0.
We are about to open-source the event recorder API since it is used in
mutter, and we want to comply with mutter's GPL license. Bumped the
minor version number from 0 to 2 since we want to be able to require
the version of the event recorder API that installs
eosmetrics/emtr-util.h and com.endlessm.Metrics.xml. Switched
maintainer.

[endlessm/eos-sdk#2043]