Add X11 (a.k.a., MIT) License.
We are open-sourcing this repository since we use it from our fork of
mutter, which is released under GNU GPLv2. The X11 License is compatible
with GNU GPLv2 but much more permissive. The X11 License allows us to
use this repository from both our closed-source and open-source code in
addition to allowing anyone else to do the same.

[endlessm/eos-sdk#2043]