Move from MIT license to LGPL 2.1.
This involved including the full license in the COPYING file
for the GPL as well as the COPYING.LESSER file for the LGPL,
both of which must be included. Also, every file referencing
the MIT license in its header or within its code-generation
sections had to be updated with the LGPL header as recommended
by the Free Software Foundation.

[endlessm/eos-sdk#2900]