Ignore production event recorder daemon in tests.
The asynchronous tests that mock out the event recorder D-Bus service
currently wait until either a method call to the production D-Bus
service or the mock D-Bus service. If the former arrives first, then a
test failure is reported because the mock D-Bus service has not yet been
called by the time we quit the main loop. Instead, we now ensure the
signal handler that quits the main loop is only called when the mock
D-Bus service is called.

[endlessm/eos-sdk#3029]