Deprecate emtr_event_id_to_name and event IDs.
eosmetrics/emtr-event-types.h is no longer the right place to add new
event IDs. Event IDs should be defined in the file in which they are
used. Otherwise we run into problems when changing the wire protocol of
those event IDs since the event ID needs to change in lock step with the
auxiliary payload. This means that emtr_event_id_to_name can't be
implemented because there's no central registry of event IDs on the
client.

[endlessm/eos-sdk#3029]