eosmetrics: Allow skipping submission when EOS_DISABLE_METRICS is set
Add an EOS_DISABLE_METRICS environment variable which, when set, causes
metrics to be quietly *not* submitted to the metrics daemon, and for the
submission code to still return successfully.

This is intended to be used from within the unit tests of other modules,
to avoid subprocesses which use eos-metrics submitting spurious metrics
to the server, or emitting warnings due to the metrics daemon not
running (and hence aborting due to G_DEBUG=fatal-warnings).

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14934