aggregate-timer: Use GCancellable to stop timer creation
When a timer is created, destroyed, and recreated very quickly,
we can end up firing multiple StartAggregateTimer() calls with
the same parameters without stopping the previous one. In these
cases, there's no way to cancel creating one timer until the
two async functions it calls are done.

Start working towards cancellable time creation by introducing
a GCancellable that cancels creating the timer (but not creating
the proxy, and stopping the timer).

Notice that this does not yet is effective since we keep the
timer object alive as long as the async operations are finished.
Next patches will fix that.