aggregate-timer: Rework refcount management
Allow EmerAggregateTimers be destroyed while the call to
StartAggregateTimer() is ongoing. This allows reaching the
finalize vfunc, which cancels the cancellable and prevents
the creation of the aggregate timer. This requires not
casting 'user_data' into EmerAggregateTimer until we're
sure the call wasn't cancelled, which is also done in this
patch.

Keep the ref/unref dance for creating the proxy for the
aggregate timer, so that it's kept alive while ongoing.