Correctly pass null payload to StartAggregateTimer
Payloads are nullable, but D-Bus (unlike GVariant) does not have a 'maybe'
type. Throughout this API, it is faked by passing a pair of parameters:

- a boolean, indicating whether the payload is conceptually non-null
- a variant, which contains the real payload if the boolean is true, or
  a dummy value <false> if not

(Side-note: a better representation IMHO would be to use a list, with a
zero-element list meaning "nothing" and a one-element list meaning "just
x". But it's 8 years too late for that.)

Previously, emtr_event_recorder_start_aggregate_timer() handled a null
payload incorrectly. It would unconditionally pass true for the
has_payload parameter, and would box the dummy value inside a second
layer of variant, <<false>>. Handle this case correctly.

https://phabricator.endlessm.com/T33232