Allow starting aggregate timer with particular UID
StartAggregateTimer() accepts a non-nullable aggregate key and a
nullable payload. The daemon internally aggregates events based on a
four-tuple of:

- date (implicit)
- UID of the calling process (currently implicit, determined by asking
  the D-Bus daemon for the UID of the caller)
- aggregate key
- payload

We currently have two users of this API:

1. Counting time an app is open for a given user. In this case the
   aggregate key and payload are identical.
2. Counting time spent logged in by each user on the system. In this
   case the aggregate key is the UID of the logged-in user, and the
   payload is null (while UIDs are not exactly unique, they are useless
   to us on the server). This event is recorded by a system-wide daemon
   running as root, so the UID in the tuple above is 0.

In the first case, the aggregate key is completely redundant. In the
second case, if we could arrange for the daemon to use the logged-in
user's UID in its internal key, rather than root (0), then we would not
need the aggregate key either. So it seems we can probably get rid of
the aggregate key and simplify the API.

One solution would be to record the session time event from within the
user session, so that the calling process's UID would be the UID we
want. I briefly looked at this and decided against it, on the basis that
it would be sad to have an extra process running in the user
session just to call a D-Bus method at the start and then hang around
until the session ends.

The alternative solution implemented here is to make the Unix UID
caller-supplied in the D-Bus API. This was the approach taken in the
older methods on this interface. Unlike those older methods, we then
optionally expose this parameter in the C API.

Having made this change, the aggregate_key is redundant and is not
present in the new API.

Neither the C API nor the D-Bus API made it into a release, so no
attempt is being made at backwards-compatibility!

https://phabricator.endlessm.com/T33252