Add Meson
This is a first stab at moving eos-metrics to Meson. The produced
library is functional, and introspection is generated. Visibility
of public symbols is managed by a new export map file that mimics
the regex used by libtool.

Documentation continues to be generated with gtk-doc.

Tests are generated individually, since each one has a different
setup and execution method.